separate replicated/shared control plane tests
