+++ tests/integration/multicluster/sharedcontrolplane/main_test.go
@@ -12,11 +12,15 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
-package multicluster
+package sharedcontrolplane
 
 import (
 	"fmt"
 	"testing"
+
+	"istio.io/istio/pkg/test/framework/components/environment/kube"
+
+	"istio.io/istio/tests/integration/multicluster"
 
 	"istio.io/istio/pkg/test/framework"
 	"istio.io/istio/pkg/test/framework/components/istio"
@@ -32,6 +36,7 @@ var (
 	pilots             []pilot.Instance
 	clusterLocalNS     namespace.Instance
 	controlPlaneValues string
+	nClusters          int
 )
 
 func TestMain(m *testing.M) {
@@ -53,6 +58,9 @@ func TestMain(m *testing.M) {
 			// Store the cluster-local namespace.
 			clusterLocalNS = ns
 
+			// Store the number of clusters so we can create the topology
+			nClusters = len(ctx.Environment().Clusters())
+
 			// Set the cluster-local namespaces in the mesh config.
 			controlPlaneValues = fmt.Sprintf(`
 values:
@@ -66,7 +74,16 @@ values:
 				ns.Name())
 			return nil
 		}).
+		Setup(kube.Setup(func(s *kube.Settings) {
+			// Make all clusters use the same control plane
+			s.ControlPlaneTopology = make(map[resource.ClusterIndex]resource.ClusterIndex)
+			primaryCluster := resource.ClusterIndex(0)
+			for i := 0; i < nClusters; i++ {
+				s.ControlPlaneTopology[resource.ClusterIndex(i)] = primaryCluster
+			}
+		})).
 		SetupOnEnv(environment.Kube, istio.Setup(&ist, func(cfg *istio.Config) {
+			cfg.IstioOperatorConfigYAML()
 			// Set the control plane values on the config.
 			cfg.ControlPlaneValues = controlPlaneValues
 		})).
@@ -85,9 +102,9 @@ values:
 }
 
 func TestMulticlusterReachability(t *testing.T) {
-	ReachabilityTest(t, pilots)
+	multicluster.ReachabilityTest(t, pilots)
 }
 
 func TestClusterLocalService(t *testing.T) {
-	ClusterLocalTest(t, clusterLocalNS, pilots)
+	multicluster.ClusterLocalTest(t, clusterLocalNS, pilots)
 }
